move new model package to controller
