+++ src/main/java/org/icgc/argo/workflow_management/controller/RunsApi.java
@@ -23,8 +23,8 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import javax.validation.Valid;
-import org.icgc.argo.workflow_management.model.wes.RunsRequest;
-import org.icgc.argo.workflow_management.model.wes.RunsResponse;
+import org.icgc.argo.workflow_management.controller.model.wes.RunsRequest;
+import org.icgc.argo.workflow_management.controller.model.wes.RunsResponse;
 import org.icgc.argo.workflow_management.exception.model.ErrorResponse;
 import org.springframework.web.bind.annotation.RequestBody;
 import reactor.core.publisher.Mono;
