+++ src/main/java/org/icgc/argo/workflow_management/controller/impl/RunsApiController.java
@@ -21,8 +21,8 @@ package org.icgc.argo.workflow_management.controller.impl;
 import javax.validation.Valid;
 import lombok.val;
 import org.icgc.argo.workflow_management.controller.RunsApi;
-import org.icgc.argo.workflow_management.model.wes.RunsRequest;
-import org.icgc.argo.workflow_management.model.wes.RunsResponse;
+import org.icgc.argo.workflow_management.controller.model.wes.RunsRequest;
+import org.icgc.argo.workflow_management.controller.model.wes.RunsResponse;
 import org.icgc.argo.workflow_management.service.WorkflowExecutionService;
 import org.icgc.argo.workflow_management.service.model.WESRunParams;
 import org.springframework.beans.factory.annotation.Autowired;
