+++ src/main/java/org/icgc/argo/workflow_management/controller/model/wes/RunsRequest.java
@@ -16,16 +16,36 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.icgc.argo.workflow_management.model.graphql;
+package org.icgc.argo.workflow_management.controller.model.wes;
 
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import io.swagger.annotations.ApiModel;
+import java.util.HashMap;
+import java.util.Map;
+import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-import org.icgc.argo.workflow_management.model.wes.RunsRequest;
 
 @Data
-@JsonNaming()
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GqlRunsRequest extends RunsRequest {
-  private GqlWorkflowEngineParams workflowEngineParams = new GqlWorkflowEngineParams();
+@ApiModel(description = "A JSON of required and optional fields to run a workflow")
+public class RunsRequest {
+  @NotBlank(message = "workflow_url is a required field!")
+  private String workflowUrl;
+
+  private Map<String, Object> workflowParams = new HashMap<>();
+  private WorkflowEngineParams workflowEngineParams = new WorkflowEngineParams();
+
+  private Map<String, Object> workflowType;
+  private String[] workflowTypeVersion;
+  private Map<String, Object> tags;
+
+  // we will not be accepting this (at least to start)
+  private String[] workflowAttachment;
 }
